+++ BE/jackdang/src/main/java/com/jackdang/service/members/MemberService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.jackdang.controller.members.dto.MemberV2Dto;
+import com.jackdang.controller.members.dto.MemberDto;
 import com.jackdang.domain.entity.members.Member;
 import com.jackdang.domain.repository.members.MemberRepository;
 
@@ -23,14 +23,14 @@ public class MemberService {
 	 *  회원 가입
 	 */
 	@Transactional
-	public Long save(MemberV2Dto memberV2Dto) {
-		validateDuplicateMember(memberV2Dto); // 중복 회원 검증
-		memberRepository.save(memberV2Dto.toEntity());
-		return memberV2Dto.getId();
+	public Long save(MemberDto memberDto) {
+		validateDuplicateMember(memberDto); // 중복 회원 검증
+		memberRepository.save(memberDto.toEntity());
+		return memberDto.getId();
 	}
 
-	private void validateDuplicateMember(MemberV2Dto memberV2Dto) {
-		List<Member> findMembers = memberRepository.findByPhone(memberV2Dto.getPhone());
+	private void validateDuplicateMember(MemberDto memberDto) {
+		List<Member> findMembers = memberRepository.findByPhone(memberDto.getPhone());
 		
 		if (!findMembers.isEmpty()) {
 			throw new IllegalStateException("이미 존재하는 회원입니다.");
@@ -41,28 +41,21 @@ public class MemberService {
 	public List<Member> findMembers() {
 		return memberRepository.findAll();
 	}
+	
 	// 회원 단건 조회
-	public MemberV2Dto findById(Long memberId) {
+	public MemberDto findById(Long memberId) {
 		Member member = memberRepository.findById(memberId).orElseThrow(() -> new IllegalArgumentException("해당 아이디가 없습니다"));
-		return new MemberV2Dto(member);
+		return new MemberDto(member);
 	}
+	
 	// 회원 단건 수정
-	public Long update(Long memberId, MemberV2Dto memberV2Dto) {
+	@Transactional
+	public Long update(Long memberId, MemberDto memberDto) {
 		Member member = memberRepository.findById(memberId).orElseThrow(() -> new IllegalArgumentException("해당 아이디가 없습니다"));
-		member.update(memberV2Dto.getAge(), memberV2Dto.getIntroduce(), memberV2Dto.getAddress(), memberV2Dto.getJob(), memberV2Dto.getSchool(), memberV2Dto.getMbti(), memberV2Dto.getLove_status(), memberV2Dto.getReligion());
+		member.update(memberDto.getPassword(), memberDto.getAge(), memberDto.getIntroduce(), memberDto.getAddress(), memberDto.getJob(), memberDto.getSchool(), memberDto.getMbti(), memberDto.getLove_status(), memberDto.getReligion());
 		return memberId;
 	};
-//	
-//	public Member findOne(Long memberId) {
-//		return memberRepository.findById(memberId).get();
-//	}
-	
-//	@Transactional
-//	public void update(Long id, String password) {
-//		Member member = memberRepository.findById(id).get();
-//		member.setPassword(password);
-//		
-//	}
+
 
 	public boolean login(String phone, String password){
 		List<Member> findMembers = memberRepository.findByPhone(phone);
